Create Analysis.py
Takes the data from collect. Using it to find Heart rate, Spo2, and makes a prediction. This is all done after the data is filtered and presented to the user. The reason this and the collection code are separate is that this program can take data sets in .csv files. Allowing us to analyze data sets from outside our own.